Update update overall rating code
