+++ shooter/helper.py
@@ -1,6 +1,9 @@
 import requests
 from bs4 import BeautifulSoup
 
+from shooter import app, db
+from shooter.models import Teacher, Rating
+
 
 def ykps_auth(username: str, password: str):
     '''
@@ -30,3 +33,19 @@ def ykps_auth(username: str, password: str):
 
     return ret, name
 
+
+def update_teacher_overall(teacher_id, new_rating, user_id):
+    '''Updates the overall score of a teacher before a new rating is inserted.
+    Currently only using simple arithmetic mean, will change in future.'''
+
+    teacher = Teacher.query.get(teacher_id)
+    current_overall = teacher.rating
+    num_ratings = Rating.query.filter_by(teacher_id=teacher_id).count()
+
+    new_rating = int(new_rating) / 2
+
+    new_overall = (num_ratings * current_overall + new_rating) / (num_ratings + 1)
+
+    teacher.rating = new_overall
+    db.session.commit()
+    return 0
